refactor and fixed copy
