+++ src/core/webview/ClineProvider.ts
@@ -1524,7 +1524,7 @@ export class ClineProvider
 			hasOpenedModeSelector: this.getGlobalState("hasOpenedModeSelector") ?? false,
 			alwaysAllowFollowupQuestions: alwaysAllowFollowupQuestions ?? false,
 			followupAutoApproveTimeoutMs: followupAutoApproveTimeoutMs ?? 60000,
-			toolExecutionTimeoutMs: toolExecutionTimeoutMs ?? 300000,
+			toolExecutionTimeoutMs: toolExecutionTimeoutMs ?? 60000,
 			timeoutFallbackEnabled: timeoutFallbackEnabled ?? false,
 		}
 	}
@@ -1638,7 +1638,7 @@ export class ClineProvider
 			terminalZshP10k: stateValues.terminalZshP10k ?? false,
 			terminalZdotdir: stateValues.terminalZdotdir ?? false,
 			terminalCompressProgressBar: stateValues.terminalCompressProgressBar ?? true,
-			toolExecutionTimeoutMs: stateValues.toolExecutionTimeoutMs ?? 300000, // 5 minutes default
+			toolExecutionTimeoutMs: stateValues.toolExecutionTimeoutMs ?? 60000,
 			timeoutFallbackEnabled: stateValues.timeoutFallbackEnabled ?? false,
 			mode: stateValues.mode ?? defaultModeSlug,
 			language: stateValues.language ?? formatLanguage(vscode.env.language),
