+++ webview-ui/src/i18n/locales/en/chat.json
@@ -230,7 +230,7 @@
 		"hasQuestion": "Roo has a question:"
 	},
 	"taskCompleted": "Task Completed",
-	"toolTimeout": "Tool Timeout",
+	"toolTimeout": "Automatic Tool Timeout: Suggesting alternatives...",
 	"error": "Error",
 	"diffError": {
 		"title": "Edit Unsuccessful"
